Add EDM UPM build script (#542)
