+++ src/Assets/Bugsnag/Editor/BugsnagEditor.EDM.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 55fc4b302697348259084ed568410d50
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
