+++ upm-tools/EDM/BugsnagAndroidDependencies.xml.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5d46d1afeed764371aaea8340dbe58cc
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
